+ mapping the NUMBER type with scale to Double schema type

+++ kafka-connect-cdc-logminer/src/main/java/com/github/jcustenborder/kafka/connect/cdc/logminer/oracle/OracleChangeBuilder.java
@@ -146,7 +146,7 @@ public class OracleChangeBuilder {
                     break;
                 case FLOAT32:
                 case FLOAT64:
-                    field = Field.create(Field.Type.FLOAT, columnValue);
+                    field = Field.create(Field.Type.DOUBLE, columnValue);
                     break;
                 default:
                     log.warn("Unsupport type: " + schema.type().name());
